-added token based auth for users -refactored tests to use tokens

+++ part4/blogs/controllers/blogs.js
@@ -1,7 +1,5 @@
 const blogRouter = require('express').Router()
-const { assignWith } = require('lodash')
 const Blog = require('../models/blog')
-const User = require('../models/user')
 
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', {
@@ -13,20 +11,38 @@ blogRouter.get('/', async (request, response) => {
 })
 
 blogRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
+    const body = request.body
+
+    const user = request.user
+    if (!user) {
+        return response.status(401).json({ error: 'Unauthorized' })
+    }
+    const blog = new Blog({
+        ...body,
+        user: user._id,
+    })
 
     if (!blog.url || !blog.title) {
         return response.status(400).json({ error: 'Missing blog url or title' })
     }
-
-    const randomUser = await User.findOne({})
-    blog.user = randomUser._id
 
     const result = await blog.save()
     return response.status(201).json(result)
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+    const user = request.user
+    if (!user) {
+        return response.status(400).json({ error: 'You are not logged in' })
+    }
+    const blogToDelete = await Blog.findById(request.params.id)
+
+    if (blogToDelete.user.toString() !== user.id.toString()) {
+        return response
+            .status(401)
+            .json({ error: 'Trying to delete another user`s content' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
